fix: make fmt
